Implement parsing
Due to the nature of logical expressions, both the parsing process and
the stmt objects are recursive. The basic idea is, when parsing a binary
expression, the two operands are parsed recursively unless they are '0',
'1', or an atomic statement. Similarly, when evaluating a binary
expression at a given set of truth values (represented by the truth
struct), the two operands are recursively evaluated before having the
operation applied to them.

Overall, the solution is pretty clean, though the parseRecursive method
may receive an update in the future to clean up the state machine
design.

Also, write some tests for the parsing code, which are all tests
currently passing.