Implement fmt.Stringer for stmts
This allows reconstructing a "pretty version" of the input string. A
single space is inserted between operators and operands, and redundant
negations and parentheses are removed; otherwise, the statement is
faithfully recreated as a string.

Also, write tests for the String method, all of which are passing.

(Removing redundant negations is not necessarily desirable, so this will
likely change in the future.)