Add basic truthtable printing functionality
The truthtables use Unicode box-drawing characters and ANSI color escape
sequences to colorize false as red and true as green by default. Both
the characters used to draw the table and whether or not to colorize the
output is cutomizable; other than that, there is no customizability,
and that might change in the future.