Reverse bit order in Truth.Val
I.e. now the 0th bit corresponds to 'z' and the 51st bit corresponds to
'A' if every letter is used.

This allows us to display each atomic in alphabetical order in a truth
table yet still have the rows count up as if they were a binary number
(i.e. {0-0, 0-1, 1-0, 1-1} order instead of {0-0, 1-0, 0-1, 1-1}, which
is what it was before), yet still be able to iterate through all the
truth value sets via incrementing Truth.Val.